+++ programs/Makefile
@@ -64,6 +64,7 @@ APPS = \
 	aes/crypt_and_hash \
 	hash/generic_sum \
 	hash/hello \
+	hash/hmac_demo \
 	pkey/dh_client \
 	pkey/dh_genprime \
 	pkey/dh_server \
@@ -87,8 +88,7 @@ APPS = \
 	psa/aead_non_psa \
 	psa/aead_psa \
 	psa/crypto_examples \
-	psa/hmac_non_psa \
-	psa/hmac_psa \
+	psa/hmac_demo \
 	psa/key_ladder_demo \
 	psa/psa_constant_names \
 	random/gen_entropy \
@@ -177,13 +177,17 @@ aes/crypt_and_hash$(EXEXT): aes/crypt_and_hash.c $(DEP)
 	echo "  CC    aes/crypt_and_hash.c"
 	$(CC) $(LOCAL_CFLAGS) $(CFLAGS) aes/crypt_and_hash.c $(LOCAL_LDFLAGS) $(LDFLAGS) -o $@
 
+hash/generic_sum$(EXEXT): hash/generic_sum.c $(DEP)
+	echo "  CC    hash/generic_sum.c"
+	$(CC) $(LOCAL_CFLAGS) $(CFLAGS) hash/generic_sum.c $(LOCAL_LDFLAGS) $(LDFLAGS) -o $@
+
 hash/hello$(EXEXT): hash/hello.c $(DEP)
 	echo "  CC    hash/hello.c"
 	$(CC) $(LOCAL_CFLAGS) $(CFLAGS) hash/hello.c       $(LOCAL_LDFLAGS) $(LDFLAGS) -o $@
 
-hash/generic_sum$(EXEXT): hash/generic_sum.c $(DEP)
-	echo "  CC    hash/generic_sum.c"
-	$(CC) $(LOCAL_CFLAGS) $(CFLAGS) hash/generic_sum.c $(LOCAL_LDFLAGS) $(LDFLAGS) -o $@
+hash/hmac_demo$(EXEXT): hash/hmac_demo.c $(DEP)
+	echo "  CC    hash/hmac_demo.c"
+	$(CC) $(LOCAL_CFLAGS) $(CFLAGS) hash/hmac_demo.c    $(LOCAL_LDFLAGS) $(LDFLAGS) -o $@
 
 pkey/dh_client$(EXEXT): pkey/dh_client.c $(DEP)
 	echo "  CC    pkey/dh_client.c"
@@ -277,13 +281,9 @@ psa/crypto_examples$(EXEXT): psa/crypto_examples.c $(DEP)
 	echo "  CC    psa/crypto_examples.c"
 	$(CC) $(LOCAL_CFLAGS) $(CFLAGS) psa/crypto_examples.c    $(LOCAL_LDFLAGS) $(LDFLAGS) -o $@
 
-psa/hmac_non_psa$(EXEXT): psa/hmac_non_psa.c $(DEP)
-	echo "  CC    psa/hmac_non_psa.c"
-	$(CC) $(LOCAL_CFLAGS) $(CFLAGS) psa/hmac_non_psa.c    $(LOCAL_LDFLAGS) $(LDFLAGS) -o $@
-
-psa/hmac_psa$(EXEXT): psa/hmac_psa.c $(DEP)
-	echo "  CC    psa/hmac_psa.c"
-	$(CC) $(LOCAL_CFLAGS) $(CFLAGS) psa/hmac_psa.c    $(LOCAL_LDFLAGS) $(LDFLAGS) -o $@
+psa/hmac_demo$(EXEXT): psa/hmac_demo.c $(DEP)
+	echo "  CC    psa/hmac_demo.c"
+	$(CC) $(LOCAL_CFLAGS) $(CFLAGS) psa/hmac_demo.c    $(LOCAL_LDFLAGS) $(LDFLAGS) -o $@
 
 psa/key_ladder_demo$(EXEXT): psa/key_ladder_demo.c $(DEP)
 	echo "  CC    psa/key_ladder_demo.c"
