+++ programs/hash/hmac_demo.c
@@ -1,11 +1,11 @@
 /**
- * PSA API multi-part HMAC demonstration.
+ * MD API multi-part HMAC demonstration.
  *
  * This programs computes the HMAC of two messages using the multi-part API.
  *
- * It comes with a companion program hmac_non_psa.c, which does the same
- * operations with the legacy MD API. The goal is that comparing the two
- * programs will help people migrating to the PSA Crypto API.
+ * This is a companion to psa/hmac_demo.c, doing the same operations with the
+ * legacy MD API. The goal is that comparing the two programs will help people
+ * migrating to the PSA Crypto API.
  *
  * When it comes to multi-part HMAC operations, the `mbedtls_md_context`
  * serves a dual purpose (1) hold the key, and (2) save progress information
@@ -13,7 +13,7 @@
  * objects: (1) a psa_key_id_t to hold the key, and (2) a psa_operation_t for
  * multi-part progress.
  *
- * This program and its companion hmac_non_psa.c illustrate this by doing the
+ * This program and its companion psa/hmac_demo.c illustrate this by doing the
  * same sequence of multi-part HMAC computation with both APIs; looking at the
  * two side by side should make the differences and similarities clear.
  */
@@ -40,7 +40,7 @@
  * standard C headers for functions we'll use here. */
 #include "mbedtls/build_info.h"
 
-#include "psa/crypto.h"
+#include "mbedtls/md.h"
 
 #include "mbedtls/platform_util.h" // for mbedtls_platform_zeroize
 
@@ -48,12 +48,10 @@
 #include <stdio.h>
 
 /* If the build options we need are not enabled, compile a placeholder. */
-#if !defined(MBEDTLS_PSA_CRYPTO_C) || \
-    defined(MBEDTLS_PSA_CRYPTO_KEY_ID_ENCODES_OWNER)
+#if !defined(MBEDTLS_MD_C)
 int main( void )
 {
-    printf( "MBEDTLS_PSA_CRYPTO_C not defined, "
-            "and/or MBEDTLS_PSA_CRYPTO_KEY_ID_ENCODES_OWNER defined\r\n" );
+    printf( "MBEDTLS_MD_C not defined\r\n" );
     return( 0 );
 }
 #else
@@ -71,7 +69,7 @@ const unsigned char msg2_part2[] = { 0x06, 0x06 };
 const unsigned char key_bytes[32] = { 0 };
 
 /* Print the contents of a buffer in hex */
-void print_buf( const char *title, uint8_t *buf, size_t len )
+void print_buf( const char *title, unsigned char *buf, size_t len )
 {
     printf( "%s:", title );
     for( size_t i = 0; i < len; i++ )
@@ -79,89 +77,68 @@ void print_buf( const char *title, uint8_t *buf, size_t len )
     printf( "\n" );
 }
 
-/* Run a PSA function and bail out if it fails. */
-#define PSA_CHECK( expr )                                       \
+/* Run an Mbed TLS function and bail out if it fails. */
+#define CHK( expr )                                             \
     do                                                          \
     {                                                           \
-        status = ( expr );                                      \
-        if( status != PSA_SUCCESS )                             \
+        ret = ( expr );                                         \
+        if( ret != 0 )                                          \
         {                                                       \
             printf( "Error %d at line %d: %s\n",                \
-                    (int) status,                               \
+                    ret,                                        \
                     __LINE__,                                   \
                     #expr );                                    \
             goto exit;                                          \
         }                                                       \
-    }                                                           \
-    while( 0 )
+    } while( 0 )
 
 /*
  * This function demonstrates computation of the HMAC of two messages using
  * the multipart API.
  */
-psa_status_t hmac_demo(void)
+int hmac_demo(void)
 {
-    psa_status_t status;
-#define ALG PSA_ALG_HMAC(PSA_ALG_SHA_256)
-    const psa_algorithm_t alg = ALG;
-    // compilers with insufficient C99 support don't accept the const variable
-    // 'alg' here, so use a macro instead in order to pacify them
-    uint8_t out[PSA_MAC_LENGTH(PSA_KEY_TYPE_HMAC, 8 * sizeof( key_bytes ), ALG)];
+    int ret;
+    const mbedtls_md_type_t alg = MBEDTLS_MD_SHA256;
+    unsigned char out[MBEDTLS_MD_MAX_SIZE]; // safe but not optimal
 
-    psa_key_attributes_t attributes = PSA_KEY_ATTRIBUTES_INIT;
-    psa_key_id_t key = 0;
+    mbedtls_md_context_t ctx;
 
-    /* prepare key */
-    psa_set_key_usage_flags( &attributes, PSA_KEY_USAGE_SIGN_MESSAGE );
-    psa_set_key_algorithm( &attributes, alg );
-    psa_set_key_type( &attributes, PSA_KEY_TYPE_HMAC );
-    psa_set_key_bits( &attributes, 8 * sizeof( key_bytes ) ); // optional
+    mbedtls_md_init( &ctx );
 
-    status = psa_import_key( &attributes, key_bytes, sizeof( key_bytes ), &key );
-    if( status != PSA_SUCCESS )
-        return( status );
-
-    /* prepare operation */
-    psa_mac_operation_t op = PSA_MAC_OPERATION_INIT;
-    size_t out_len = 0;
+    /* prepare context and load key */
+    // the last argument to setup is 1 to enable HMAC (not just hashing)
+    CHK( mbedtls_md_setup( &ctx, mbedtls_md_info_from_type( alg ), 1 ) );
+    CHK( mbedtls_md_hmac_starts( &ctx, key_bytes, sizeof( key_bytes ) ) );
 
     /* compute HMAC(key, msg1_part1 | msg1_part2) */
-    PSA_CHECK( psa_mac_sign_setup( &op, key, alg ) );
-    PSA_CHECK( psa_mac_update( &op, msg1_part1, sizeof( msg1_part1 ) ) );
-    PSA_CHECK( psa_mac_update( &op, msg1_part2, sizeof( msg1_part2 ) ) );
-    PSA_CHECK( psa_mac_sign_finish( &op, out, sizeof( out ), &out_len ) );
+    CHK( mbedtls_md_hmac_update( &ctx, msg1_part1, sizeof( msg1_part1 ) ) );
+    CHK( mbedtls_md_hmac_update( &ctx, msg1_part2, sizeof( msg1_part2 ) ) );
+    CHK( mbedtls_md_hmac_finish( &ctx, out ) );
     print_buf( "msg1", out, sizeof( out ) );
 
     /* compute HMAC(key, msg2_part1 | msg2_part2) */
-    PSA_CHECK( psa_mac_sign_setup( &op, key, alg ) );
-    PSA_CHECK( psa_mac_update( &op, msg2_part1, sizeof( msg2_part1 ) ) );
-    PSA_CHECK( psa_mac_update( &op, msg2_part2, sizeof( msg2_part2 ) ) );
-    PSA_CHECK( psa_mac_sign_finish( &op, out, sizeof( out ), &out_len ) );
+    CHK( mbedtls_md_hmac_reset( &ctx ) ); // prepare for new operation
+    CHK( mbedtls_md_hmac_update( &ctx, msg2_part1, sizeof( msg2_part1 ) ) );
+    CHK( mbedtls_md_hmac_update( &ctx, msg2_part2, sizeof( msg2_part2 ) ) );
+    CHK( mbedtls_md_hmac_finish( &ctx, out ) );
     print_buf( "msg2", out, sizeof( out ) );
 
 exit:
-    psa_mac_abort( &op ); // needed on error, harmless on success
-    psa_destroy_key( key );
+    mbedtls_md_free( &ctx );
     mbedtls_platform_zeroize( out, sizeof( out ) );
 
-    return( status );
+    return( ret );
 }
 
 int main(void)
 {
-    psa_status_t status = PSA_SUCCESS;
+    int ret;
 
-    /* Initialize the PSA crypto library. */
-    PSA_CHECK( psa_crypto_init( ) );
-
-    /* Run the demo */
-    PSA_CHECK( hmac_demo() );
-
-    /* Deinitialize the PSA crypto library. */
-    mbedtls_psa_crypto_free( );
+    CHK( hmac_demo() );
 
 exit:
-    return( status == PSA_SUCCESS ? EXIT_SUCCESS : EXIT_FAILURE );
+    return( ret == 0 ? EXIT_SUCCESS : EXIT_FAILURE );
 }
 
 #endif
