+++ programs/psa/CMakeLists.txt
@@ -2,8 +2,7 @@ set(executables
     aead_non_psa
     aead_psa
     crypto_examples
-    hmac_non_psa
-    hmac_psa
+    hmac_demo
     key_ladder_demo
     psa_constant_names
 )
